requirements.txt
appnope==0.1.3 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "darwin"
asgiref==3.6.0 ; python_version >= "3.10" and python_version < "4.0"
asttokens==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
backcall==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
beautifulsoup4==4.11.1 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
decorator==5.1.1 ; python_version >= "3.10" and python_version < "4.0"
dj-database-url==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
django-bootstrap4==22.3 ; python_version >= "3.10" and python_version < "4.0"
django-extensions==3.2.1 ; python_version >= "3.10" and python_version < "4.0"
django-filter==22.1 ; python_version >= "3.10" and python_version < "4.0"
django==4.1.5 ; python_version >= "3.10" and python_version < "4.0"
executing==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
ipython==8.8.0 ; python_version >= "3.10" and python_version < "4.0"
jedi==0.18.2 ; python_version >= "3.10" and python_version < "4.0"
matplotlib-inline==0.1.6 ; python_version >= "3.10" and python_version < "4.0"
parso==0.8.3 ; python_version >= "3.10" and python_version < "4.0"
pexpect==4.8.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "win32"
pickleshare==0.7.5 ; python_version >= "3.10" and python_version < "4.0"
prompt-toolkit==3.0.36 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.5 ; python_version >= "3.10" and python_version < "4.0"
ptyprocess==0.7.0 ; python_version >= "3.10" and python_version < "4.0" and sys_platform != "win32"
pure-eval==0.2.2 ; python_version >= "3.10" and python_version < "4.0"
pygments==2.14.0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.10" and python_version < "4.0"
setuptools==66.1.1 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
soupsieve==2.3.2.post1 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
stack-data==0.6.2 ; python_version >= "3.10" and python_version < "4.0"
traitlets==5.8.1 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2022.7 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
wcwidth==0.2.6 ; python_version >= "3.10" and python_version < "4.0"
whitenoise==6.3.0 ; python_version >= "3.10" and python_version < "4.0"
